aws-scripts/cdk_multi_value_ssm_parameter/requirements.txt
aws-cdk-lib==2.59.0
constructs>=10.0.0,<11.0.0
boto3==1.26.47
